+++ src/entities/bonuses/store/index.ts
@@ -1,0 +1,2 @@
+export * from './gifts';
+export * from './giftsV2';