+++ src/entities/promotions/store/index.ts
@@ -1,0 +1,2 @@
+export * from "./promo";
+export * from "./tournaments";