+++ CHANGELOG.md
@@ -1,3 +1,69 @@
+## [7.30.0](https://github.com/upstars-global/front-platform-core/compare/v7.29.0...v7.30.0) (2026-04-20)
+
+### 🚀 Features
+
+* **FP-5067:** add additional check for data
+ ([d01aabb](https://github.com/upstars-global/front-platform-core/commit/d01aabb47a04932423cdd32b5d6ebacb0e327eb4))
+
+
+* **FP-5067:** add current-game
+ ([30b7492](https://github.com/upstars-global/front-platform-core/commit/30b7492759abf338894e8cf052c91ce153aa7ec5))
+
+
+* **FP-5067:** add games
+ ([68eef8e](https://github.com/upstars-global/front-platform-core/commit/68eef8ea16db9dac39072a19927236645148784c))
+
+
+* **FP-5067:** rename useCurrentGame to useCurrentGameStore
+ ([65c6f6f](https://github.com/upstars-global/front-platform-core/commit/65c6f6f5df30aef531f09cd4b8a0884ecf718faa))
+
+
+* **FP-5067:** simplify response handling in LOAD_GAME_BY_SLUG
+ ([8e2d2e7](https://github.com/upstars-global/front-platform-core/commit/8e2d2e743ab440b7b88c0a7f3a3f218487ddbfcb))
+
+## [7.29.0](https://github.com/upstars-global/front-platform-core/compare/v7.28.0...v7.29.0) (2026-04-20)
+
+### 🚀 Features
+
+* **web-chat:** add API for loading JSON web token
+ ([2b3fd1d](https://github.com/upstars-global/front-platform-core/commit/2b3fd1dc95acb6894c30c23253a700574275f22b))
+
+## [7.28.0](https://github.com/upstars-global/front-platform-core/compare/v7.27.2...v7.28.0) (2026-04-16)
+
+### 🚀 Features
+
+* **FP-5076:** added promotions
+ ([fda638c](https://github.com/upstars-global/front-platform-core/commit/fda638c6a030f16013ce9e53a17693ec382f1836))
+
+
+* **FP-5076:** import fixes
+ ([c5b4f49](https://github.com/upstars-global/front-platform-core/commit/c5b4f498739dad5476911db2ebe1aa0abce888e3))
+
+
+* **FP-5076:** return bonuses to old place
+ ([82f3b33](https://github.com/upstars-global/front-platform-core/commit/82f3b33d643a06630f239da61538fe409f15b21e))
+
+## [7.27.2](https://github.com/upstars-global/front-platform-core/compare/v7.27.1...v7.27.2) (2026-04-16)
+
+### 🔧 Maintenance
+
+* **FP-4842:** mark isSelfExclusionEnabled as deprecated
+ ([0479e3b](https://github.com/upstars-global/front-platform-core/commit/0479e3b6552f0b716c833085a487aad2f0414f6a))
+
+## [7.27.1](https://github.com/upstars-global/front-platform-core/compare/v7.27.0...v7.27.1) (2026-04-16)
+
+### 🔨 Refactoring
+
+* **FP-5148:** remove email replacement from useFillWithUserData
+ ([09480db](https://github.com/upstars-global/front-platform-core/commit/09480db6416982383ee87b37f0d469822c4fb85e))
+
+## [7.27.0](https://github.com/upstars-global/front-platform-core/compare/v7.26.0...v7.27.0) (2026-04-14)
+
+### 🚀 Features
+
+* **FP-3177:** added bonuses
+ ([9cb66d8](https://github.com/upstars-global/front-platform-core/commit/9cb66d8b5c6ff3ecdba1c62292e4103cd2e4bcc4))
+
 ## [7.26.0](https://github.com/upstars-global/front-platform-core/compare/v7.25.0...v7.26.0) (2026-04-09)
 
 ### 🚀 Features